Adds notification token persistence
Implements logic to persist Expo push tokens associated with user install IDs to the database.

Adds background modes to `infoPlist` for location, fetch, and remote-notification.

Displays an alert if notification permissions are denied, prompting the user to enable them in settings.

Also generates and stores a unique installation UUID upon the first app launch, which is then used as an identifier to associate the push token with the user.